moved subSectorInfo into its own component
