+++ client/src/modules/EmissionPieChart.css
@@ -3,6 +3,7 @@
   padding: 10px;
   display: flex;
   justify-content: center;
+  height: 500px;
 }
 .container-chart-subsector {
   flex: 1;
@@ -11,7 +12,8 @@
   list-style-type: none;
   margin: 0;
   padding: 0;
-  min-width: 200px;
+  min-width: 400px;
+  text-align: left;
 }
 .recharts-wrapper {
   flex: 1;
